TransformTools : Don't edit read-only EditScopes
This bug caused a strange behaviour with locked EditScope nodes. The transform handles were initially moveable, allowing the user to start a drag. But as soon as the drag started, we created the edit and immediately realised that it was read-only, and so made the handles non-moveable. The result : the object didn't move, but an unwanted edit was still created in the EditScope.