SConstruct : Cache inkscape builds
Graphics builds are slow due to the overhead of running the inkscape
process for each export. We now declare each image to scons as a
target. This allows us to take advantage of the build cache when
jumping between previously build commits with different svg hashes.

Export images are declared in a sidecar file, listing svg object
ids. As we now directly reference ids, we no longer need to use
the `forExport:` prefix. This commit removes the prefix from the
existing project svg files.