Use `universal_newlines = True` when getting subprocess output
This argument is ostensibly about Unix vs Windows line endings, but more importantly it uses the default encoding - as specified by `locale.getpreferredencoding()` - to convert from bytes to strings. In Python 3 we could use the much more appropriately named `text` argument, but we want something compatible with both 2 and 3, and this is it.