TransformTool : Disable editing of locations with zero scale components
This results in a plethora of errors later on when attempting to
calculate related spaces, etc. This ensures handles are never drawn,
avoiding a plethora of draw/tool interaction related errors.
In the `0.56.x` series, this does result in a somewhat unhelpful error
message in the Viewer, but is preferable to a crash. `0.57.x` provides
a mechanism that will allow us to present the true error.

TODO: On merge forward, fix confusing error messages, add tests.