ShadingEngine : Fix string lookups
Although `RendererServices::get_userdata()` passes strings to OSL as `const char *`, there's an implicit assumption that they have already been uniquefied using `OIIO::ustring`, because string comparisons in OSL compare by identity rather than contents. We've been bitten by this already in OSLExpressionEngine - see 963bf1966f267ff8f0ee9a6edf05544f8c0eed65.

We thought we had test coverage for this already, but in fact the test was broken. In a rare moment of helpfulness, Python 3 pointed this out by limiting the scope of `x` to the first string comprehension rather than leaking it out as Python 2 would. We had been using this leaked variable in the second loop all this time, not noticing that it meant we were only testing a single index in the array.