IECoreScenePreview : Allow optional message handler to be passed to renderer
In many situations, we wish to present render information to the user.
Presently, most renderers output to the terminal directly This commit
allows renderer owners to provide an IECore::MessageHandler when creating a renderer.
If supplied, supporting renderers will/should re-direct any messages to this handler.

We opted to provide the handler during construction (rather than set
via accessors), as some real-world renderers require their message
handlers to be present during initialisation of the render context.