TransformTool : Work around Tool/View lifetime issues
I think the long-term solution here is to have Views own Tools via the regular GraphComponent parenting mechanism, probably stored in a `View["tools"]` container. This would be in keeping with our general plan of turning the whole UI layout into a GraphComponent hierarchy so that it can be serialised using the same mechanism we use for graphs. This isn't as straightforward as just sorting out the parenting though. Parenting implies reparenting, and the reason we forced a single View to be used for the lifetime of a tool was to avoid the complexity that reparenting would add.