PointsType : Change base class from Deformer to ObjectProcessor
The cost of the bounds propagation done by Deformer isn't justified by the small improvement in accuracy it provides. The only time PointsType affects a bound is when changing to and from `type == "patch"`. In this case there are two sources of bounds changes :

- `PointsPrimitive::bound()` uses the diagonal of a patch of the specified width, rather than using width directly. This error was worth worrying about when we rendered everything as a procedural (where incorrect bounds would clip the procedural in the rendered image), but is no longer a genuine concern.
- `PointsPrimitive::bound()` considers a "patchaspectratio" primitive variable that isn't supported in any of our current renderer backends (it was a 3Delight thing, back before NSI).

Ignoring both the above is consistent with our existing decision to ignore changes to the "width" primvar in our various `Deformer::adjustBounds()` overloads.