Gaffer : Remove `lazyImport()` method
We were only using this for IECoreGL in Viewer.py, but that was completely ineffective because we were loading non-lazily elsewhere. Our other use for lazy loading was loading certain Qt modules, but that is no longer possible because we're using Qt.py.

The implementation we were using doesn't work in Python 3, because `reload()` is gone. It would be possible to reimplement differently, and Python 3 even has an `importlib.LazyLoader()` ( though they don't really recommend it). But since we're not actually losing any benefit by removing, I've opted to just simplify things for now.