SpreadsheetUI : Add "enabled" switch to cell edit window
This is more convenient than the "Enable/Disable Cell" menu items, and allows full access to the plug for keyframing/expressions etc. It felt wrong that NameValuePlug/TweakPlug cells could be enabled/disabled in the popup window but other plugs couldn't, and this makes that experience a bit more uniform.