EventLoop : Fix call to `traceback.format_exc()`
The first argument is optional, and should be a numeric limit, not the exception object. The exception object is implicitly the current exception. This has always been the case, but in Python 2 we got away with it and in Python 3 we got this exception :

```
  File "/disk1/john/dev/build/gafferPython3/lib/python3.7/traceback.py", line 337, in extract
    if limit >= 0:
TypeError: '>=' not supported between instances of 'RuntimeError' and 'int'
```