Add CI using Github Actions
This currently omits GafferArnold, until we figure out what to do about the login secret (the GitHub rules are different to the Azure ones). But it has the major benefit of caching builds. My intention is to use this to set up Python 3 testing, as we did for Cortex, then to make the setup more comprehensive and use it to replace Azure.