BackdropNodeGadget : Avoid creating unnecessary plugs
We were adding a `__uiBound` plug in the gadget constructor, but during copy/paste, the Gadget is constructed when the node is first added to the script, and then any dynamic plugs are added afterwards. This meant a duplicate `__uiBound1` plug being added as the paste operation completes.

Although the `__uiBound` plug is private, it doesn't seem totally unlikely that people may be relying on setting it directly, so we provide the new `setBound()` and `getBound()` methods as a legitimate means of achieving the same thing. There's a reasonable argument for adding a public `Backdrop::boundsPlug()` instead, but dealing with bounds without references to Gadgets is problematic, because without them you don't know how big a node is. So for now I've gone with fixing the bug in the existing scheme.