AnimationEditor : Improve _AnimationPathFilter performance
When selecting all nodes, the previous code was quadratic in the number of nodes, as it looped through the selection once per node. When selecting 2000 nodes, the new approach reduces the filtering runtime to less than 4% of the original. In truth, this code is still slower than would be ideal, and should probably be implemented in C++.