Context : Allow removal of canceller during copy construction
I had deliberately made it impossible to remove a canceller from a context, because doing so would be against the wishes of whoever launched a computation with a canceller in it. But we now have a use case where it is necessary to capture a copy of a context without the canceller.

This means of specifying cancellers was always a bit awkward, and this doesn't really improve that. It's particularly bothersome that the copy constructor is intended for when you want an independent long-lived context that owns its own data (and EditableScope is the lightweight non-owning alternative for temporary contexts). This isn't reflected in the Canceller semantics at all, where the lifetime management is left in the hands of the caller.

A better long-term solution may be to further the split between Context and EditableScope, and only allow cancellers to exist in EditableScopes (or their future equivalent). The Context class (or its future equivalent)  would then be about long-term ownership where cancellation is irrelevant. And we could make it utterly impossible to remove a canceller from an EditableScope, so cancellation is guaranteed for entire process trees.