Dispatcher : Fix bug handling multiple connected Switches/ContextProcessors
Refactored the special handling for these nodes into a separate function, which uses recursion to deal with multiple connected instances in a row.

We don't really have a standard policy for how to return multiple values from one function, so I've experimented with `std::tuple()`. This is nice in that it'd map to Python naturally, where output parameters often don't make sense. The `tie()` stuff at the call site is a bit faffy, but becomes unnecessary when we finally make it to `C++17`, where you can do `auto [ source, context ] = computedSource()` instead. It's potentially confusing not giving names to the types, as a struct would, but in this case the types are distinct enough for it not to matter. Structs may still be the clearest in some ways, but they introduce a lot of namespace pollution with small types all over the place, and they're more faff for binding into Python. The jury is out...