IECorePreview : Add Messages data type
The Messages data type provides cheap-to-copy storage for log messages.

We debated the pros/cons of a mutable container with change notifications,
vs a cheap-to-copy immutable container. It seemed that in many use cases,
the container would be frequently updated and be shared between many
UI elements (eg, render messages shown in log windows, message count
widgets, etc...).

In the mutable case, the overhead of coordinating locking between all
clients seemed problematic, especially as we expect high ingest rates.
As such, we favoured the immutable option as it minimises the
synchronisation overhead at the expense of copy cost.

This initial implementation takes a simple, but reasonably effective
approach to copy optimisation that works well for the current use cases.
It is however somewhat tuned to count/rate (via bucketSize). A more
general purpose implementation could improve this situation significantly.