StyleSheet : Extend font stack for monospace
On mac, the `monospace` font-family didn't actually results in a
monospace font being used. Calls to QFontDatabase also seem to return
the wrong font (which is perhaps why).

This borrows the GitHub monospace system font stack, which also ensures
we don't just end up with Courier unless we really have to.