GafferUI : Fix button highlight if disabled during hover
The button would remain in a highlighted state when highlightOnHover
is true, and someone disables the button in a click signal callback.

We don't receive leave events once disabled, so have no opportunity to
reset the appearance for when we are enabled again. As such, we
forcibly unset highlighting in setEnabled if it was due to a hover.