MessageWidget : Support larger number of messages
Motivated by the needs to efficiently display thousands of messages
from process output, this commit refactors the widget to be based on
a QTableView instead of a MultiLineTextWidget.

The table view implementation provides a seemingly efficient view of
a window of data from a larger set, so works well with the cheap-to-copy
Gaffer.Messages container, that allows large message sets to be easily
added to the widget.

This commit removes the deprecated `appendMessage` method. Individual
messages can be added via `messageHandler().handle()` if required.

This commit adds the following features to the existing widget:

 - Optional 'log' style presentation (message-per-row)
 - Search