ImageReader config : Add default colorspace for RAW files
Without this the function simply fails. I'm basing the choice of `linear` on what I see in the OIIO source code, and some utterly inexpert eyeballing of images from https://rawsamples.ch. The OIIO source is best understood from this commit :

https://github.com/OpenImageIO/oiio/commit/35d00c048afa3f3620240bee1b77a6ede8d5e142

We are still using an older OIIO, which puts the code we are using on the red side of the diff. I _think_ that means OIIO is requesting output in linear space with sRGB primaries. And I _think_ it means that when we upgrade OIIO it will instead be requesting sRGB primaries _and_ response. When we upgrade we should revisit this and probably expose the "raw:ColorSpace" option via the ImageReader.