Filter : Refactor to allow for full dirty propagation
Filters are unconventional because their computes use ScenePlugs that do not belong to them, but are instead passed in via the Context. This has always been fairly nasty, but it allows the same filter to be used for multiple nodes without a tangle of automatic connection management behind the scenes.

Because the ScenePlugs aren't owned by the filters, `Filter::affects()` is not called automatically when the scene was dirtied. We previously worked around this by calling `Filter::sceneAffectsMatch()` from `FilteredSceneProcessor::affects()` and then manually dirtying the input FilterPlug when it returned true. This successfully propagated dirtiness back to the FilteredSceneProcessor, but it didn't give the Filter an opportunity to dirty any of its own internal plugs which may have depended on the scene. This caused subtle dirty propagation bugs that were hidden by the lack of observers for the internal plugs and the brute force clearing of the _entire_ hash cache when _any_ plug was dirtied.

We now provide full dirty propagation for all plugs belonging to a Filter by replacing calls to `Filter::sceneAffectsMatch()` with more general calls to `Filter::affects()`.