Shader : Rationalise dirty propagation
The true target of computations for the Shader node is the `outAttributes()` plug, so our approach is now centred around this. The new `affectsAttributes()` virtual method mirrors the `affectsFoo()` methods that accompany `hashFoo()` and `computeFoo()` methods throughout the API, and gives derived classes a mechanism whereby they can dirty the `outAttributes()` plug. We then simply dirty `outPlug()` automatically whenever `outAttributesPlug()` is dirtied, to propagate dirtiness to downstream ShaderPlugs.

This allows us to fix bugs in ArnoldDisplacement and ArnoldCameraShaders whereby `outAttributes()` wasn't being dirtied when when the displacement/shader plugs were modified.