InteractiveRenderTest : Ensure UI thread call handler
As InteractiveRender currently and does a bunch of work in
`plugDirtied`, any exceptions thrown updating the renderer
end up being propagated to the `errorSignal()` (which no-one is
currently listening to). This can make debugging test failures
somewhat hard.

Many tests were manually scoping a handler. This change simply moves
this up to the base class `run` method so we know know one is
always available. It can be accessed through `self._uiThreadHandler`.