InteractiveRenderTest : Catch errorSignal messages during tests
Previously, any errors raised via `errorSignal` were lost in
test cases. These are now bridged to the default IECore.MessageHandler,
which will in turn cause appropriate test failures.

This is backwards-compatible with existing tests, and will be enabled
for common tests with no changes, but renderer-specific test cases
may need to be updated.