Expression : Fix obscure dirty propagation problem
This was exposed by `OSLExpressionEngineTest.testDefaultExpression` when switching between an expression driving a float and one driving an int. In this case the _type_ of the output plugs actually affects the engine's execution result. We could in theory implement `affects()` so that the addition of the new output plug dirties the `__execute` plug, and use `Plug::Flags::AcceptsDependencyCycles` to break the resulting cycle, but I don't think it buys much over the simpler hack.