SeedsTest : Fix broken test
This test wasn't actually testing anything - it was supposed to ensure that a new density primvar took affect,
but was checking that the point count is <= to the previous count ... it's equal if it's failing and nothing is
happening.  Once corrected, this test fails unless the hash cache is disabled.