Catalogue : Support renaming of images via promoted plugs
The `images` plug maintains a list of Images as shown in the
catalogue. The name of its child plugs defines the names of each
image. When the `images` plug is promoted, standard plug code takes
care of propagating child additions/removals to the catalogue's own
`images` plug.

The catalogue considers the `catalogue:imageName` context variable, to
override the current `imageIndex` to compute outputs from the
requested image instead.

This was achieved by caching the name->index for all images
(presumably for performance reasons). The cache was rebuilt whenever
a child `Image` was added/removed or renamed. This worked fine for
standard Catalogue nodes, but there is no rename propagation from the
children of promoted plugs. This meant that renaming a child of the
promoted `images` counterpart was not tracked by the connected
Catalogue.

There didn't seem to be a reliable route to tracking the `nameChanged`
signal from children of `images->source()`. But it seems that
`mappingPlug` is relatively infrequently accessed anyway.

This commit simplifies the implementation by removing the mapping
cache and preforming a direct lookup in the source's children
instead.

As this fix is targeted as a patch, leave the plug/accessors in the
short term until this is merged forward to an ABI break compatible
release.

Fixes #3815.