DependencyNodeBinding : Track initialisation via a custom metaclass
This allows us to query if the Python `__init__()` method has finished executing for a particular DependencyNode instance. We don't use this query for anything in this commit - that will happen in the next commit.

If this metaclass technique works out in practice, we think we could generalise it into RefCountedBinding, so that instead of setting a flag, we would automatically call a `virtual postConstructor()` method following the construction of any RefCounted instance from Python. This would be far more flexible and with far wider applications.

We would also want to support automatic `postConstructor()` calls when constructing via C++, but there we don't have metaclasses to help us out. Still, we have been considering protecting all RefCounted constructors anyway, so people can't do `new RefCountedThing` and fail to assign the result to an `intrusive_ptr`. Instead we would force construction via a `makeRefCounted()` method in the spirit of `std::make_shared`, and this could be responsible for calling the `postConstructor()`.