Plug : Propagate dirtiness for all parent changes
Previously we only did this for dynamic plugs, to protect `DependencyNode::affects()` from being called with partially constructed nodes. We now implement the protection in DownstreamIterator instead, assuming that nodes without a refcount are still being constructed. This appears to work well for C++ nodes, and allows us to remove some workarounds in AttributeProcessor and ObjectProcessor. But it isn't sufficient for Python nodes, where the `__init__` methods run _after_ the refcount has been incremented. For Python, we instead use the new `DependencyNodeWrapper::initialised()` query to avoid calling `affects()` overrides in Python before `__init__()` has completed. It is a pity to have different mechanisms for each, but in the future we may be able to use a single `DependencyNode::postConstructor()` for both, using the ideas documented in the preceding commit.