SpreadsheetUI/SceneViewInspectorUI : Don't focus disabled widgets
When displaying a popup editor, we attempt to focus the first
text field and select its contents for editing. If the widget is
disabled, the selection is still visible, but the widget does not
have keyboard focus.

Should the widget subsequently be enabled (eg: turning on
a tweak). This results in the field _looking_ like the text is
selected and pressing a key will interact with it. However, as the
widget doesn't have focus, nothing happens.

This change avoids selecting/assigning focus if the target widget
is disabled.

A potential future extension of this would be a allow classes such
as TweakValuePlugWidget to apply this focus-and-select behaviour to
their value fields when a tweak is enabled via the switch in the UI
if the field is editable.