EditScopeAlgo : Add 'readOnlyReason' methods
It is often necessary to determine why an edit scope edit is read-only,
or, why it may not be possible to add an edit to an edit scope. These
helper functions make it easier for UIs to surface this information
to users, and avoid inadvertently modifying read-only
scopes/processor/rows plugs, etc.