TransformTool : Include negative scaling in `orientedTransform()`
This fixes the direction of the handles shown by the TranslateTool when negative scaling is present. Before they were not pointing in the same direction as the axes of the advertised space, which was confusing.

This also allows us to simplify the implementation of `ScaleTool::updateHandles()`, since it has been manually including the scaling itself the whole time.