SceneNode : Fix `hashChildBounds()` bug
By assigning the reduction to `h` rather than appending, we were obliterating the part of the hash that reflects the plug itself and the TypeId of the node, encouraging collisions with other plugs and/or nodes.

As an aside, this does make me realise that initialising the hash with the virtual `typeId()` is somewhat pessimistic when particular computes are implemented on a base class rather than on the most derived class. In theory we could get better cache sharing if we initialised with the `typeId()` of the class that implements the compute for that specific plug. This could perhaps be achieved by passing an explicit TypeId to the base class `hash()` method. I'm unsure if this would have enough real-world benefit to warrant the extra complexity though.