GafferScene : Fix `childBoundsPlug()` dependency bugs
When `childBoundsPlug()` was introduced as a faster cached alternative to `unionOfTransformedChildBounds()`, I neglected to update nodes to declare the dependency on the new plug. That was OKish until we introduced the improved hash cache invalidation, at which point it meant that we could lookup stale results from the cache. The fix is simple - just declare the dependencies accurately. For many of the nodes I've also taken the time to refactor `affects()` so that there is a single conditional per output plug, making it much easier to verify that it is in agreement with the contents of hash/compute.