Plug : Fix bug in dependency cycle handling
By throwing an exception and then turning it into a message, rather than just emitting a message, we were aborting the `depth_first_search()` before we had called `Plug::dirty()` and `Node::plugDirtiedSignal()` for all plugs in the dirty graph. This is particularly important now that plugs need to be dirtied in order to invalidate their entries in the hash cache.