ShaderPlug : Relax `acceptsInput()` restrictions
These were tightened too much in #2321 via c6dfc6c79ff58b8c79831d40f09343c09682beb4. That PR was about tightening up `acceptsInput()` for specific plug types (ShaderPlug, TaskPlug, FilterPlug and ClosurePlug) which were introduced to replace the use of more generic plug types (Plug and IntPlug). TaskPlug, FilterPlug and ClosurePlug should never accept new inputs from their former generic types, except to support the loading of old scripts, hence we tightened the checks in cases where loading wasn't being performed. It was a mistake to do the same for ShaderPlug though. ShaderPlug is different in that it can accept inputs not only from plugs of its type, but also from any suitable parameter plug being output from a Shader. The acceptance of intermediate inputs from Dots etc therefore needs to be allowed at any time, not just when loading old scripts.

This improves some interactions in the UI, allowing a Dot output to be connected to a ShaderAssignment before a Shader is connected to the Dot input. It also fixes a crash when destroying a node network with a Dot feeding through a Switch into a ShaderAssignment, as demonstrated in `ShaderAssignmentTest.testSwitchGraphDestruction()`.