GitHub Actions : Publish build as artefact
Non-release builds are now published as artefacts of the parent CI run.

In order to avoid naming collisions, this introduces a `variant` matrix
variable that should be used to uniquely distinguish each build. This
is used in place of the previous implicit `platform` name component.

NOTE: GitHub has a variety of limitations surrounding the download of
artefacts.

  - They are double compressed. Whatever is published is always re-wrapped
    in a zip file for download, even if it is already a compressed archive.
    Attempting to upload an uncompressed Gaffer build takes 40+ minutes.

  - You _must_ be logged in to download them. This is due to GitHub's
    security model, and can't be avoided even for public repos.

Release builds are unaffected by these limitations.

We surveyed the current use of PR builds and based on the nature of regular
users, decided this was an acceptable compromise. It avoids the
cost/complication of artefacts via Azure blob uploads.