GitHub Actions : Validate builds and publish releases
Now we have Arnold support in GitHub CI Builds, we can validate and publish
release tag builds.

We choose to use a more generic approach to this, rather than the official
Actions available in the marketplace (`publish-release-asset`). At the
time of writing, this required an upload url, rather than a release ID,
so is better suited to workflows that make user of their companion
`create-release` action.