GitHub Actions : Add problemMatchers to CI steps
Azure provided us with comprehensive test coverage integration out the box.

Sadly, Actions is a little lacking in this aspect. It does however provide
a `problemMatcher` mechanism. It allows you to generate annotations based
on output from CI steps. At the time of writing, this seems to be the only
way to surface any meaningful failure information higher up in the CI UI.

In practice, this doesn't work so well.

 - Matchers are really about annotating source file lines, and so any
   meaningful information about file/line number provided by the matcher
   results in links to the relevant file in the commit. There is no way
   to link back to the output lines in the log. This is less than helpful
   for test coverage, as not only may the failure be entirely unrelated
   to the files that changed in the commit, but you actually want to see
   the specific test failure output.

 - Matchers have basic support for multi-line matching, but its not suited
   to python tracebacks, or the `unittest` format, that both have a
   variable number of lines to ignore between the the initial match and
   a subsequent match that contains pertinent information. There is also
   no way to aggregate information from multiple lines in the message.

All in all, this commit attempts to make the best of this situation, it's
not a great improvement, but it does at least allow you see at the top
level CI page, which tests failed, or what is missing from validation.

`validateRelease.py` output needed to be updated to produce more meaningful
text in the annotation.