Docs : Build sphinx docs with system python
In order to build the docs, sphinx et. al. need to be installed. Now we
build Gaffer with multiple python versions, attempting to use modules
installed in the host python env is fraught with problems. It requires
the host environment to have the pre-requisites installed for all matching
python versions, these paths communicated to the docs build process
independently of the version of python that is running scons.

This commit moves to building the docs using the host python environment,
such that the docs toolchain now only needs to be installed for this
one environment, regardless of the Python version in the target Gaffer
build.