CollectScenes : Refactor context management
This splits the single SceneScope class into two, with the SourceScope providing what we need when evaluating globals, and the SourcePathScope providing what we need when evaluating locations.

This also changes the constructors so that we pass `CollectScenes *` to the SourcePathScope so it can access any plugs it wants. This is something we typically avoid doing because it obfuscates the dependencies and makes it harder to implement `affects()` correctly. But the new `SourcePathScope::affectedBy()` method hopefully makes things clear enough. The motivation is that we're going to need additional information in SourcePathScope soon, and the constructor was already a bit unwieldy.