Plug : Fix cache invalidation bug for plugs with AcceptsDependencyCycles
We use `depth_first_search()` to call `Plug::dirty()` and emit `plugDirtiedSignal()` for dirtied plugs in order, upstream plugs first, children before parents. But where a plug willingly participates in a dependency cycle as with the Loop node, we cannot guarantee the "upstream first" property for all observers. What constitutes "upstream" in such a cycle is dependent on your point of view. This was OK until Gaffer 0.58, since when `Plug::dirty()` has no longer invalidated the entire cache but now selectively invalidates the entries for individual plugs. This could lead to an observer triggering a compute from `plugDirtiedSignal()` before `dirty()` had been called on all "upstream" plugs. This would then cause the compute to see invalid cache entries for a subset of upstream plugs.

The fix is to decouple the calls to `Plug::dirty()` and `plugDirtiedSignal()`. We now call `Plug::dirty()` when first adding the plug to the graph of dirty plugs, so it is guaranteed to have been called before `plugDirtiedSignal()` is emitted for _any_ plug.

As an aside, I now wonder if the "upstream first" property of `plugDirtiedSignal()` is worth preserving at all, particularly as we are making increasing use of `Plug::AcceptsDependencyCycles`. As we optimise other things, dirty propagation is beginning to factor into interactive profiles, and it might be that relaxing the ordering requirements could provide some scope for improvements. The "children before parents" order must be preserved though.

Fixes #3911.