GafferSceneUITest : Add CropWindowTool tests
The CropWindowTool is in need of a refactor, its logic is somewhat complex
and coupled with the render mechanism. The approach taken in TransformTool
or similar would be beneficial once refactor time permits.

As such, these tests rely on constructing a view, and interrogating the
tools status messages, not ideal, but hopefully a useful stop-gap.

NOTE: When merging forward to 0.58, we also have access to the plugs used
buy the tool. These tests should be updated to include checks for those
accessors too.