OCIO config : Simplify setup of default display transform
Rather than using signals and slots to manually ferry value updates across from the preferences to the transform, we can express the same thing far more easily via a single call to `setInput()`. This also has the benefit that we no longer need to store `__defaultDisplayTransforms`, so the transforms are disposed of correctly when their true owners are done with them.

This approach also fixes a potential crash in the UVView, where DisplayTransforms are currently stored such that they can't participate in cancellation. Now they receive an input from the preferences, the preferences will trigger cancellation.