GafferImage Tests : Update for OIIO 2.2+
OIIO 2.2 includes a 'nuke-lanczos6' filters that the reference images do not have. Rather than regenerate the ref images and cause test failures with older OIIO, I've chosen to hardcode the list of expected filters and add a print out about the untested filters.

OIIO 2.2 also no longer considers tiffs to have a dataWindow, and will not read or write that metadata. Note non-OIIO apps like RV do not read this metadata, and our reference tiffs load as though they are smaller in RV. See https://github.com/OpenImageIO/oiio/pull/2521 for an explanation. I've chosen not to replace the ref images but instead ignore the dataWindow on the tiff tests.