IECoreArnoldPreview::ShaderNetworkAlgo : Fix for OSL 1.11+
For Arnold, we don't seem to need the same fix as in `GafferOSL::ShadingEngine`. Perhaps the original todo was valid in this case because Arnold is doing the work for us.

Note I'm not entirely positive this is correct. All GafferArnoldTests are paassing, but I've yet to use this interactively and actually see it shading correctly. Regarding the changes I've made to `GafferArnoldTest.IECoreArnoldPreviewTest.RendererTest.testOSLShaderComponentConnections`, the shaders in the .ass file are as follows:

```
osl
{
 name shader:38e598c1eeef2402dd8ef9b497d1d21d
 shadername "Maths/MixColor"
 param_a.r shader:38e598c1eeef2402dd8ef9b497d1d21d:source:out.b
 param_a.g shader:38e598c1eeef2402dd8ef9b497d1d21d:source:out.r
 param_a.b shader:38e598c1eeef2402dd8ef9b497d1d21d:source:out.g
}

osl
{
 name shader:38e598c1eeef2402dd8ef9b497d1d21d:source:out
 declare output constant STRING
 output "out"
 shadername "Maths/MixColor"
}
```